refactor: move LessonDetails, LessonElemenLink, LessonItemContainer nad LessonResourceGraphics
