+++ src/components/TeacherComponents/LessonDetails/LessonDetails.test.tsx
@@ -2,7 +2,6 @@ import LessonDetails from "./LessonDetails";
 
 import renderWithTheme from "@/__tests__/__helpers__/renderWithTheme";
 
-
 describe("LessonDetails component", () => {
   const keyLearningPoints = [{ keyLearningPoint: "test" }];
   const commonMisconceptions = [
