+++ src/components/TeacherComponents/LessonDetails/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./LessonDetails";
