+++ src/components/TeacherComponents/LessonItemContainer/index.tsx
@@ -1,0 +1,1 @@
+export { LessonItemContainer } from "./LessonItemContainer";
