+++ src/components/TeacherComponents/LessonItemContainerLink/index.tsx
@@ -1,0 +1,1 @@
+export { LessonItemContainerLink } from "./LessonItemContainerLink";
