+++ src/components/TeacherComponents/LessonResourceGraphics/LessonResourceGraphics.stories.tsx
