+++ src/components/TeacherComponents/LessonResourceGraphics/index.tsx
