+++ src/components/TeacherComponents/LessonResourceGraphicsItem/LessonResourceGraphicsItem.tsx
