+++ src/components/TeacherComponents/LessonResourceGraphicsItem/index.tsx
@@ -1,0 +1,2 @@
+export { default } from "./LessonResourceGraphicsItem";
+export type { LessonResourceGraphicsItemProps } from "./LessonResourceGraphicsItem";
